requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
DateTime==5.4
flask-login==0.6.3
flask-restful==0.3.10